Modify to use Keyword Search + 4.10 compatable
Change the search method to use keyword search.
This allows individual tollerences to be configured by each trigger
phrase.
The change to use keyword search, required a small modification to
Pocketsphinx.
The modifications to the Pocketsphinx source are contained in the _other
folder.
I have also added a Blueprint that shows how to use the speech
recognition plugin. (SpeechRecognitionCharacter.uasset)